(refactor): move tests to test/, test-utils/ to test/config/
- have been putting tests separately from source nowadays in other of
  my libraries, so standardize around that

- put fixtures just in test/ (later could be in test/fixtures/), leave
  config/ solely for test config type things
